projects: fmcjesdadc1: use new iio generic app
Replace custom fmcjesdadc1 iio app with the generic one.
