+++ projects/fmcjesdadc1/src/app/fmcjesdadc1.c
@@ -62,7 +62,8 @@
 #include "demux_spi.h"
 
 #ifdef IIO_SUPPORT
-#include "app_iio.h"
+#include "iio_app.h"
+#include "iio_axi_adc.h"
 #endif
 
 /***************************************************************************//**
@@ -542,8 +543,10 @@ int main(void)
 #ifdef IIO_SUPPORT
 	printf("The board accepts libiio clients connections through the serial backend.\n");
 
-	struct iio_axi_adc_init_param iio_axi_adc_init_par;
-	iio_axi_adc_init_par = (struct iio_axi_adc_init_param) {
+	struct iio_axi_adc_desc *iio_axi_adc_desc;
+	struct iio_device *adc_dev_desc;
+
+	struct iio_axi_adc_init_param iio_axi_adc_init_par = {
 		.rx_adc = ad9250_core,
 		.rx_dmac = ad9250_dmac,
 #ifndef PLATFORM_MB
@@ -552,7 +555,23 @@ int main(void)
 #endif
 	};
 
-	return iio_server_init(&iio_axi_adc_init_par);
+	status = iio_axi_adc_init(&iio_axi_adc_desc, &iio_axi_adc_init_par);
+	if (status < 0)
+		return status;
+
+	iio_axi_adc_get_dev_descriptor(iio_axi_adc_desc, &adc_dev_desc);
+
+	struct iio_data_buffer g_read_buff = {
+		.buff = (void *)ADC_DDR_BASEADDR,
+		.size = 0xFFFFFFFF,
+	};
+
+	struct iio_app_device devices[] = {
+		IIO_APP_DEVICE("axi_adc", iio_axi_adc_desc, adc_dev_desc,
+			       &g_read_buff, NULL),
+	};
+
+	return iio_app_run(devices, ARRAY_SIZE(devices));
 
 #endif
