+++ projects/fmcjesdadc1/src/devices/adi_hal/parameters.h
@@ -66,6 +66,8 @@
 #define ADC_DDR_BASEADDR			(XPAR_DDR_MEM_BASEADDR + 0x800000)
 #endif
 
+#define UART_BAUDRATE	            		115200
+
 #define RX_CORE_BASEADDR			XPAR_AXI_AD9250_CORE_ADC_TPL_CORE_BASEADDR
 
 #define RX_DMA_BASEADDR			XPAR_AXI_AD9250_DMA_BASEADDR
